add use cmd enter #40
