+++ app/common/hooks/use-cmd-enter.tsx
@@ -1,0 +1,16 @@
+import { useEffect } from "react";
+
+/** keyboard shortcut */
+export default function useCmdEnter(onSubmit: () => Promise<void> | void) {
+  useEffect(() => {
+    const submitShortcut = (e: KeyboardEvent) => {
+      if (e.key === "Enter" && (e.metaKey || e.ctrlKey)) {
+        onSubmit();
+      }
+    };
+    document.addEventListener("keydown", submitShortcut);
+    return () => {
+      document.removeEventListener("keydown", submitShortcut);
+    };
+  }, [onSubmit]);
+}
